Compute the value of the geodesic curve connecting two points on the unit sphere for a given time point t

This function computes points along the geodesic connecting two points p and q on the unit sphere.

Usage

geodesic_lower(t, p, q, a, b)

Arguments

t

Time parameter for the geodesic path.

p

Numeric vector representing the starting point.

q

Numeric vector representing the ending point.

a

Start time parameter.

b

End time parameter.

Value

Numeric vector representing a point along the geodesic path.

Examples

geodesic_lower(0.5, c(1,0,0), c(0,1,0), 0, 1)
#> [1] 0.7071068 0.7071068 0.0000000